after PR review: added UIButton extension
